Authorizing Actions

+++ users/app/Http/Controllers/Controller.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Query\Builder;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Cache;
 use Laravel\Lumen\Routing\Controller as BaseController;
 
@@ -17,6 +19,7 @@ abstract class Controller extends BaseController
      */
     protected Model $model;
     protected string $api;
+
     /**
      * @return JsonResponse
      */
@@ -39,7 +42,22 @@ abstract class Controller extends BaseController
      *
      * @return JsonResponse
      */
-    abstract public function store(Request $request): JsonResponse;
+    public function store(Request $request): JsonResponse
+    {
+        try {
+            $this->authorize('create', Auth::user());
+            Cache::forget($this->api);
+            $data = $request->all();
+            /**
+             * @var Model $model
+             */
+            $model = new $this->model($data);
+            $model->save();
+            return response()->json($model, 201);
+        } catch (AuthorizationException $e) {
+            return response()->json(['message' => $e->getMessage()], 403);
+        }
+    }
 
     /**
      * @param string $id
@@ -48,8 +66,13 @@ abstract class Controller extends BaseController
      */
     public function show(string $id): JsonResponse
     {
-        $userType = $this->model->find($id);
-        return response()->json($userType, 200);
+        try {
+            $this->authorize('view', Auth::user());
+            $userType = $this->model->find($id);
+            return response()->json($userType, 200);
+        } catch (AuthorizationException $e) {
+            return response()->json(['message' => $e->getMessage()], 403);
+        }
     }
 
     /**
@@ -60,12 +83,17 @@ abstract class Controller extends BaseController
      */
     public function update(Request $request, string $id): JsonResponse
     {
-        Cache::forget($this->api);
+        try {
+            $this->authorize('update', Auth::user());
+            Cache::forget($this->api);
 
-        /** @var Model $data */
-        $data = $this->model->find($id);
-        $data->update($request->all());
-        return response()->json($data, 200);
+            /** @var Model $data */
+            $data = $this->model->find($id);
+            $data->update($request->all());
+            return response()->json($data, 200);
+        } catch (AuthorizationException $e) {
+            return response()->json(['message' => $e->getMessage()], 403);
+        }
     }
 
     /**
@@ -76,10 +104,15 @@ abstract class Controller extends BaseController
      */
     public function destroy(string $id): JsonResponse
     {
-        Cache::forget($this->api);
-        /** @var Model $data */
-        $data = $this->model->find($id);
-        $data->delete();
-        return response()->json(null, 204);
+        try {
+            $this->authorize('delete', Auth::user());
+            Cache::forget($this->api);
+            /** @var Model $data */
+            $data = $this->model->find($id);
+            $data->delete();
+            return response()->json(null, 204);
+        } catch (AuthorizationException $e) {
+            return response()->json(['message' => $e->getMessage()], 403);
+        }
     }
 }
